refactor: Rename package from "domain" to "models"
