+++ internal/models/service.go
@@ -1,4 +1,4 @@
-package domain
+package models
 
 import (
 	"fmt"
